exercise 3 updated

+++ exercise3.java
@@ -1,7 +1,7 @@
 public class exercise3 {
 
     public static void main(String[] args) {
-        int x = 124;
+        float x = 124.0f;
         x = (x < 100) ? (x + 50) : (x / 20);
         System.out.println(x);
     }
